Move ClosureWrappedConfigureAction to core-utils
