+++ subprojects/core-utils/src/test/groovy/dev/nokee/utils/ClosureWrappedConfigureActionTest.java
@@ -13,11 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.nokee.platform.base.internal;
+package dev.nokee.utils;
 
 import com.google.common.testing.EqualsTester;
 import com.google.common.testing.NullPointerTester;
-import dev.nokee.utils.ClosureTestUtils;
 import lombok.val;
 import org.junit.jupiter.api.Test;
