+++ subprojects/platform-base/src/test/groovy/dev/nokee/platform/base/internal/ViewAdapterTest.java
@@ -17,10 +17,7 @@ package dev.nokee.platform.base.internal;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
-import dev.nokee.utils.ActionTestUtils;
-import dev.nokee.utils.ClosureTestUtils;
-import dev.nokee.utils.SpecTestUtils;
-import dev.nokee.utils.TransformerTestUtils;
+import dev.nokee.utils.*;
 import groovy.lang.Closure;
 import lombok.val;
 import org.gradle.api.Action;
